breadcrumb footer, multilevel pages in sanic

+++ gyro.py
@@ -18,13 +18,6 @@ async def get_search_index(request):
     return response.redirect('/_static/index.js')
 
 
-@app.route('/_title_list')
-async def get_title_list(request):
-    titles = glob.glob(os.path.join('pages', '*.md'))
-    titles = [os.path.splitext(os.path.basename(t))[0] for t in titles]
-    return response.json(titles)
-
-
 @app.route('/')
 async def get_index(request):
     with open(os.path.join('sanic', 'index.html')) as inf:
@@ -32,9 +25,11 @@ async def get_index(request):
     return response.html(content)
 
 
-@app.route('/<word:[A-z0-9]+>', methods=['GET'])
+@app.route('/<word:path>', methods=['GET'])
 async def get_wikiword(request, word='index'):
     word = word.lower()
+    if word.startswith('_'):
+        abort(403)
     content = ''
     path = os.path.join('pages', word + '.md')
     if os.path.isfile(path):
@@ -43,21 +38,29 @@ async def get_wikiword(request, word='index'):
     return response.text(content)
 
 
-@app.route('/<word:[A-z0-9]*>', methods=['POST'])
+@app.route('/<word:path>', methods=['POST'])
 @app.route('/', methods=['POST'])
 async def post_wikiword(request, word='index'):
     word = word.lower()
+    if word.startswith('_'):
+        abort(403)
     path = os.path.join('pages', word + '.md')
     markdown = request.form['markdown'][0]
+
+    out_dir = os.path.dirname(path)
+    if not os.path.isdir(out_dir):
+        os.makedirs(out_dir)
     with open(path, mode='w', encoding='utf-8') as outf:
         outf.write(markdown)
         reindex_site()  # FIXME: this can be made more efficient.
         return response.json({'message': 'Saved'})
 
-@app.route('/<word:[A-z0-9]*>', methods=['DELETE'])
+@app.route('/<word:path>', methods=['DELETE'])
 @app.route('/', methods=['DELETE'])
 async def del_wikiword(request, word='index'):
     word = word.lower()
+    if word.startswith('_'):
+        abort(403)
     path = os.path.join('pages', word + '.md')
     if os.path.isfile(path):
         os.remove(path)
@@ -66,14 +69,21 @@ async def del_wikiword(request, word='index'):
 
 def reindex_site():
     docs = []
-    for doc in glob.glob(os.path.join('pages', '*.md')):
-        docname = os.path.splitext(os.path.basename(doc))[0]
+    for doc in glob.glob(os.path.join('pages', '**', '*.md'), recursive=True):
+        docname = os.path.splitext('/'.join(doc.split(os.sep)[1:]))[0]
         with open(doc, encoding='utf8') as inf:
             text = inf.read()
             docs.append({'name': docname, 'text': text})
     index_path = os.path.join('_static', 'index.js')
     with open(index_path, mode='w', encoding='utf-8') as outf:
         json.dump(docs, outf)
+
+
+@app.route('/_title_list')
+async def get_title_list(request):
+    titles = glob.glob(os.path.join('pages', '**', '*.md'), recursive=True)
+    titles = [os.path.splitext('/'.join(t.split(os.sep)[1:]))[0] for t in titles]
+    return response.json(titles)
 
 
 if __name__ == '__main__':
